requirements.txt
locustio
pyzmq

